chore: more caches in constructor; rename [TECH-532]
