+++ dhis-2/dhis-services/dhis-service-core/src/main/java/org/hisp/dhis/datavalue/DefaultAggregateAccessManager.java
@@ -33,20 +33,15 @@ import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
-import java.util.concurrent.TimeUnit;
-
-import javax.annotation.PostConstruct;
 
 import org.hisp.dhis.cache.Cache;
-import org.hisp.dhis.cache.SimpleCacheBuilder;
+import org.hisp.dhis.cache.CacheProvider;
 import org.hisp.dhis.category.CategoryOption;
 import org.hisp.dhis.category.CategoryOptionCombo;
-import org.hisp.dhis.commons.util.SystemUtils;
 import org.hisp.dhis.dataelement.DataElementOperand;
 import org.hisp.dhis.dataset.DataSet;
 import org.hisp.dhis.security.acl.AclService;
 import org.hisp.dhis.user.User;
-import org.springframework.core.env.Environment;
 import org.springframework.stereotype.Service;
 
 /**
@@ -56,36 +51,23 @@ import org.springframework.stereotype.Service;
 public class DefaultAggregateAccessManager
     implements AggregateAccessManager
 {
-    private static Cache<List<String>> CAN_DATA_WRITE_COC_CACHE;
+    private final Cache<List<String>> canDataWriteCocCache;
 
     private final AclService aclService;
 
-    private final Environment env;
-
-    public DefaultAggregateAccessManager( AclService aclService, Environment env )
+    @SuppressWarnings( { "rawtypes", "unchecked" } )
+    public DefaultAggregateAccessManager( AclService aclService, CacheProvider cacheProvider )
     {
         checkNotNull( aclService );
-        checkNotNull( env );
+        checkNotNull( cacheProvider );
 
         this.aclService = aclService;
-        this.env = env;
+        this.canDataWriteCocCache = (Cache) cacheProvider.createCanDataWriteCocCache( List.class );
     }
 
     // ---------------------------------------------------------------------
     // AggregateAccessManager implementation
     // ---------------------------------------------------------------------
-
-    @PostConstruct
-    public void init()
-    {
-        CAN_DATA_WRITE_COC_CACHE = new SimpleCacheBuilder<List<String>>()
-            .forRegion( "canDataWriteCocCache" )
-            .expireAfterWrite( 3, TimeUnit.HOURS )
-            .withInitialCapacity( 1000 )
-            .forceInMemory()
-            .withMaximumSize( SystemUtils.isTestRun( env.getActiveProfiles() ) ? 0 : 10000 )
-            .build();
-    }
 
     @Override
     public List<String> canRead( User user, DataValue dataValue )
@@ -187,7 +169,7 @@ public class DefaultAggregateAccessManager
     {
         String cacheKey = user.getUid() + "-" + optionCombo.getUid();
 
-        return CAN_DATA_WRITE_COC_CACHE.get( cacheKey, key -> canWrite( user, optionCombo ) ).orElse( null );
+        return canDataWriteCocCache.get( cacheKey, key -> canWrite( user, optionCombo ) ).orElse( null );
     }
 
     @Override
