+++ dhis-2/dhis-services/dhis-service-core/src/test/java/org/hisp/dhis/query/QueryParserTest.java
@@ -34,7 +34,7 @@ import java.util.Arrays;
 import java.util.List;
 
 import org.hisp.dhis.IntegrationTestBase;
-import org.hisp.dhis.cache.CacheContext;
+import org.hisp.dhis.cache.CacheProvider;
 import org.hisp.dhis.common.IdentifiableObjectManager;
 import org.hisp.dhis.configuration.ConfigurationService;
 import org.hisp.dhis.dataelement.DataElement;
@@ -89,7 +89,7 @@ public class QueryParserTest
     private UserSettingService userSettingService;
 
     @Autowired
-    private CacheContext cacheContext;
+    private CacheProvider cacheProvider;
 
     @Override
     public boolean emptyDatabaseAfterTest()
@@ -106,7 +106,7 @@ public class QueryParserTest
         user.addOrganisationUnit( orgUnitA );
         CurrentUserService currentUserService = new MockCurrentUserService( user );
         this.organisationUnitService = new DefaultOrganisationUnitService( organisationUnitStore, dataSetService,
-            organisationUnitLevelStore, currentUserService, configurationService, userSettingService, cacheContext );
+            organisationUnitLevelStore, currentUserService, configurationService, userSettingService, cacheProvider );
         organisationUnitService.addOrganisationUnit( orgUnitA );
         identifiableObjectManager.save( orgUnitA );
         queryParser = new DefaultJpaQueryParser( schemaService, currentUserService,
