+++ dhis-2/dhis-support/dhis-support-system/src/main/java/org/hisp/dhis/cache/CacheBuilderProvider.java
@@ -1,0 +1,58 @@
+/*
+ * Copyright (c) 2004-2021, University of Oslo
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ * Redistributions of source code must retain the above copyright notice, this
+ * list of conditions and the following disclaimer.
+ *
+ * Redistributions in binary form must reproduce the above copyright notice,
+ * this list of conditions and the following disclaimer in the documentation
+ * and/or other materials provided with the distribution.
+ * Neither the name of the HISP project nor the names of its contributors may
+ * be used to endorse or promote products derived from this software without
+ * specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+ * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+ * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
+ * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+ * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
+ * ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+ * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+package org.hisp.dhis.cache;
+
+import java.util.Map;
+
+/**
+ * Provides cache builder to build instances.
+ *
+ * @author Ameen Mohamed
+ */
+public interface CacheBuilderProvider
+{
+    /**
+     * Creates a new {@link ExtendedCacheBuilder} that can be used to build a
+     * cache that stores the valueType specified.
+     *
+     * @param valueType The class type of values to be stored in cache.
+     * @return A cache builder instance for the specified value type. Returns a
+     *         {@link ExtendedCacheBuilder}.
+     */
+    <V> CacheBuilder<V> newCacheBuilder( Class<V> valueType );
+
+    /**
+     * Creates a new {@link ExtendedCacheBuilder} that can be used to build a
+     * cache that stores the Map of keyType and valueType specified.
+     *
+     * @param valueType The class type of values to be stored in cache.
+     * @return A cache builder instance for the specified value type. Returns a
+     *         {@link ExtendedCacheBuilder}.
+     */
+    <K, V> ExtendedCacheBuilder<Map<K, V>> newCacheBuilder( Class<K> keyType, Class<V> valueType );
+}
